requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
libsql
pandas
httpx
email-validator
jinja2
